amsynth_vst: Use idle slice instead of thread for UI handling
Thanks Justin Frankel for great instructions!

On Windows, using a thread to handle UI is not a good idea. The better
and stabler way is to draw UI on idle event (effEditIdle).

By making use of idle event, actually there's no need to create a single
thread or even a process to run UI, and UIs can still be shown
gracefully in parallel with DSP backend. This is amazing!

But there's a side effect: Frame rate will limit to 30fps on WINE, and
approximately 25~28fps on Linux.