Future Dev 02: Make combo button into a standalone API
A combo button is similar to combo, but replaced dropdown with context
menu, and label with a standard button.

As I've used up to 6 combo buttons, it's time to make them into an ImGui
API so that I can reuse the code, and clarify my source code.

The API is in vendor/imgui_addons/anclark/imgui_extra_button.cpp.