editor_pane: Test GLFW's capture mouse callback
I add a callback when moving cursor in or out of GLFW window, to toggle
ImGui's io.wantCaptureMouse. This may help prevent unnecessary mouse
behaviors when interacting with host window.

As a GLFW callback must be a static function, I added a global
ImguiEditor pointer to the current instance, as well as a getContext()
method.

But actually ImGui itself will automatically set io.wantCaptureMouse
without explicitly setting it by user callbacks. This is ImGui's default
behavior.