Future Dev 08.1: Fix redirection of user preset on another hosts
Not all hosts will set program number to 0 when loading user preset.

For example, REAPER for Windows does, but REAPER for Linux don't.
When I modify a factory program and save it as a user preset, the next
time I load it, it will set program number to that factory program.

So another polyfill is needed. All we should do is to set up a switch on
effSetChunk. Host calls this opcode when loading user presets, and won't
call on factory programs. We can easily determine which kind of program
host loads.

NOTE: Factory program equals "built-in program".